refactor: break up `goose-cli/src/cli.rs` module
