+++ crates/goose-cli/src/cli/mod.rs
@@ -1,0 +1,12 @@
+pub mod main;
+pub mod session;
+pub mod telemetry;
+pub mod utils;
+
+pub use main::{cli, BenchCommand, CliProviderVariant};
+pub use session::{extract_identifier, Identifier};
+pub use telemetry::{
+    extract_tool_usage_from_session, log_if_telemetry_disabled, track_command_execution,
+    track_recipe_execution, track_session_execution,
+};
+pub use utils::{parse_key_val, InputConfig};
